initial attempt at tsv of all inputs
Tried to make a TSV of all the types of outputs (matches, english matches without spanish, no match for input1, no match for input2)

It works to a certain degree but is still outputting the 'English' 'Spanish' headers.  I also noticed that is happening in the html.  And I can't sort the full output.  Need to investigate further.